project creation

+++ skillrat/usermanagement/src/main/java/com/hlt/usermanagement/dto/AddressDTO.java
@@ -4,8 +4,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Getter
 @Setter
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -18,8 +16,8 @@ public class AddressDTO {
     private String city;
     private String state;
     private String country;
-    private BigDecimal latitude;
-    private BigDecimal longitude;
+    private Double latitude;
+    private Double longitude;
     private String postalCode;
     private Long userId;
     private Boolean isDefault;
