+++ skillrat/usermanagement/src/main/java/com/hlt/usermanagement/dto/B2BUnitDTO.java
@@ -6,7 +6,6 @@ import lombok.Getter;
 import lombok.Setter;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Set;
@@ -24,9 +23,9 @@ public class B2BUnitDTO {
 
 	private boolean enabled;
 
-	private BigDecimal businessLatitude;
+	private Double businessLatitude;
 
-	private BigDecimal businessLongitude;
+	private Double businessLongitude;
 
 	private String categoryName;
