+++ skillrat/usermanagement/src/main/java/com/hlt/usermanagement/model/AddressModel.java
@@ -4,8 +4,6 @@ import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
-
 @Entity
 @Table(name = "user_address")
 @Getter
@@ -37,10 +35,10 @@ public class AddressModel {
     private String postalCode;
 
     @Column(name = "latitude")
-    private BigDecimal latitude;
+    private Double latitude;
 
     @Column(name = "longitude")
-    private BigDecimal longitude;
+    private Double longitude;
 
     @Column(name = "is_default")
     private Boolean isDefault;
